Assign LWNode id to Issues when exporting them
